rename sparse.* back to embedding.*
